+++ node/src/integration-test-slow/kotlin/net/corda/node/services/statemachine/StatemachineGeneralErrorHandlingTest.kt
@@ -1,73 +1,22 @@
 package net.corda.node.services.statemachine
 
-import co.paralleluniverse.fibers.Suspendable
 import net.corda.client.rpc.CordaRPCClient
-import net.corda.core.flows.FlowLogic
-import net.corda.core.flows.FlowSession
-import net.corda.core.flows.InitiatedBy
-import net.corda.core.flows.InitiatingFlow
-import net.corda.core.flows.ReceiveFinalityFlow
-import net.corda.core.flows.StartableByRPC
-import net.corda.core.identity.CordaX500Name
-import net.corda.core.identity.Party
-import net.corda.core.internal.ResolveTransactionsFlow
-import net.corda.core.internal.list
-import net.corda.core.internal.readAllLines
 import net.corda.core.messaging.startFlow
-import net.corda.core.messaging.startTrackedFlow
-import net.corda.core.node.AppServiceHub
-import net.corda.core.node.services.CordaService
-import net.corda.core.serialization.CordaSerializable
-import net.corda.core.serialization.SingletonSerializeAsToken
-import net.corda.core.utilities.OpaqueBytes
-import net.corda.core.utilities.ProgressTracker
 import net.corda.core.utilities.getOrThrow
 import net.corda.core.utilities.seconds
-import net.corda.core.utilities.unwrap
-import net.corda.finance.DOLLARS
-import net.corda.finance.flows.CashIssueAndPaymentFlow
-import net.corda.node.services.Permissions
-import net.corda.node.services.api.ServiceHubInternal
 import net.corda.node.services.messaging.DeduplicationHandler
 import net.corda.node.services.statemachine.transitions.TopLevelTransition
 import net.corda.testing.core.ALICE_NAME
 import net.corda.testing.core.CHARLIE_NAME
-import net.corda.testing.core.DUMMY_NOTARY_NAME
 import net.corda.testing.core.singleIdentity
-import net.corda.testing.driver.DriverDSL
-import net.corda.testing.driver.DriverParameters
-import net.corda.testing.driver.NodeHandle
-import net.corda.testing.driver.NodeParameters
-import net.corda.testing.driver.driver
-import net.corda.testing.node.NotarySpec
-import net.corda.testing.node.TestCordapp
-import net.corda.testing.node.User
-import net.corda.testing.node.internal.FINANCE_CORDAPPS
-import net.corda.testing.node.internal.InternalDriverDSL
-import org.jboss.byteman.agent.submit.ScriptText
-import org.jboss.byteman.agent.submit.Submit
-import org.junit.Before
 import org.junit.Ignore
 import org.junit.Test
-import java.time.Duration
-import java.time.temporal.ChronoUnit
 import java.util.concurrent.TimeoutException
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
-import kotlin.test.assertTrue
 
 @Suppress("MaxLineLength") // Byteman rules cannot be easily wrapped
-class StatemachineErrorHandlingTest {
-
-    companion object {
-        val rpcUser = User("user1", "test", permissions = setOf(Permissions.all()))
-        var counter = 0
-    }
-
-    @Before
-    fun setup() {
-        counter = 0
-    }
+class StatemachineGeneralErrorHandlingTest : StatemachineErrorHandlingTest() {
 
     /**
      * Throws an exception when performing an [Action.SendInitial] action.
@@ -127,11 +76,11 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
             assertFailsWith<TimeoutException> {
-                aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                        30.seconds
+                aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                    30.seconds
                 )
             }
 
@@ -140,12 +89,12 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(3, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(1, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = aliceClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(3, discharge)
             assertEquals(1, observation)
             assertEquals(1, aliceClient.stateMachinesSnapshot().size)
             // 1 for the errored flow kept for observation and another for GetNumberOfCheckpointsFlow
-            assertEquals(2, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(2, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 
@@ -207,10 +156,10 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
-            aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                    30.seconds
+            aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                30.seconds
             )
 
             val output = getBytemanOutput(alice)
@@ -218,12 +167,12 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(3, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(0, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = aliceClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(3, discharge)
             assertEquals(0, observation)
             assertEquals(0, aliceClient.stateMachinesSnapshot().size)
             // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(1, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 
@@ -287,10 +236,10 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
-            aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                    30.seconds
+            aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                30.seconds
             )
 
             val output = getBytemanOutput(alice)
@@ -298,12 +247,12 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(0, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(0, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = aliceClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(0, discharge)
             assertEquals(0, observation)
             assertEquals(0, aliceClient.stateMachinesSnapshot().size)
             // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(1, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 
@@ -372,10 +321,10 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
-            aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                    30.seconds
+            aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                30.seconds
             )
 
             val output = getBytemanOutput(alice)
@@ -383,12 +332,12 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(3, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(0, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = aliceClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(3, discharge)
             assertEquals(0, observation)
             assertEquals(0, aliceClient.stateMachinesSnapshot().size)
             // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(1, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 
@@ -459,11 +408,11 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
             assertFailsWith<TimeoutException> {
-                aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                        30.seconds
+                aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                    30.seconds
                 )
             }
 
@@ -472,12 +421,12 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(3, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(1, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = aliceClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(3, discharge)
             assertEquals(1, observation)
             assertEquals(1, aliceClient.stateMachinesSnapshot().size)
             // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(1, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 
@@ -562,10 +511,10 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
-            aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                    30.seconds
+            aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                30.seconds
             )
 
             val output = getBytemanOutput(alice)
@@ -573,12 +522,12 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(3, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(0, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = aliceClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(3, discharge)
             assertEquals(0, observation)
             assertEquals(0, aliceClient.stateMachinesSnapshot().size)
             // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(1, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 
@@ -651,10 +600,10 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
-            aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                    30.seconds
+            aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                30.seconds
             )
 
             val output = getBytemanOutput(alice)
@@ -662,12 +611,12 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(3, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(0, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = aliceClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(3, discharge)
             assertEquals(0, observation)
             assertEquals(0, aliceClient.stateMachinesSnapshot().size)
             // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(1, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 
@@ -747,11 +696,11 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
             assertFailsWith<TimeoutException> {
-                aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                        30.seconds
+                aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                    30.seconds
                 )
             }
 
@@ -760,12 +709,12 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(1, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(0, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = aliceClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(1, discharge)
             assertEquals(1, observation)
             assertEquals(1, aliceClient.stateMachinesSnapshot().size)
             // 1 for the errored flow kept for observation and another for GetNumberOfCheckpointsFlow
-            assertEquals(2, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(2, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 
@@ -847,10 +796,10 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
-            aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                    30.seconds
+            aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                30.seconds
             )
 
             val output = getBytemanOutput(alice)
@@ -858,12 +807,12 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(1, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(0, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = aliceClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(1, discharge)
             assertEquals(0, observation)
             assertEquals(0, aliceClient.stateMachinesSnapshot().size)
             // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(1, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 
@@ -931,11 +880,11 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
             assertFailsWith<TimeoutException> {
-                aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                        30.seconds
+                aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                    30.seconds
                 )
             }
 
@@ -944,12 +893,12 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(1, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(0, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = aliceClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(1, discharge)
             assertEquals(1, observation)
             assertEquals(1, aliceClient.stateMachinesSnapshot().size)
             // 1 for the errored flow kept for observation and another for GetNumberOfCheckpointsFlow
-            assertEquals(2, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(2, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 
@@ -1023,11 +972,11 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
             assertFailsWith<TimeoutException> {
-                aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                        30.seconds
+                aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                    30.seconds
                 )
             }
 
@@ -1036,12 +985,12 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(3, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(1, output.filter { it.contains("Byteman test - not my speciality") }.size)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = aliceClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(3, discharge)
             assertEquals(1, observation)
             assertEquals(1, aliceClient.stateMachinesSnapshot().size)
             // 1 for errored flow and 1 for GetNumberOfCheckpointsFlow
-            assertEquals(2, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(2, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 
@@ -1111,12 +1060,12 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
             val charlieClient =
-                    CordaRPCClient(charlie.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(charlie.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
-            aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                    30.seconds
+            aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                30.seconds
             )
 
             val output = getBytemanOutput(charlie)
@@ -1124,15 +1073,15 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(3, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(0, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = charlieClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = charlieClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(3, discharge)
             assertEquals(0, observation)
             assertEquals(0, aliceClient.stateMachinesSnapshot().size)
             assertEquals(0, charlieClient.stateMachinesSnapshot().size)
             // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(1, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
             // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, charlieClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(1, charlieClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 
@@ -1206,13 +1155,13 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
             val charlieClient =
-                    CordaRPCClient(charlie.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(charlie.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
             assertFailsWith<TimeoutException> {
-                aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                        30.seconds
+                aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                    30.seconds
                 )
             }
 
@@ -1221,17 +1170,17 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(3, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(1, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = charlieClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = charlieClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(3, discharge)
             assertEquals(1, observation)
             assertEquals(1, aliceClient.stateMachinesSnapshot().size)
             assertEquals(1, charlieClient.stateMachinesSnapshot().size)
             // 1 for the flow that is waiting for the errored counterparty flow to finish and 1 for GetNumberOfCheckpointsFlow
-            assertEquals(2, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(2, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
             // 1 for GetNumberOfCheckpointsFlow
             // the checkpoint is not persisted since it kept failing the original checkpoint commit
             // the flow will recover since artemis will keep the events and replay them on node restart
-            assertEquals(1, charlieClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(1, charlieClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 
@@ -1305,12 +1254,12 @@ class StatemachineErrorHandlingTest {
             submitBytemanRules(rules)
 
             val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
             val charlieClient =
-                    CordaRPCClient(charlie.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
+                CordaRPCClient(charlie.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
 
-            aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
-                    30.seconds
+            aliceClient.startFlow(StatemachineErrorHandlingTest::SendAMessageFlow, charlie.nodeInfo.singleIdentity()).returnValue.getOrThrow(
+                30.seconds
             )
 
             val output = getBytemanOutput(charlie)
@@ -1318,750 +1267,15 @@ class StatemachineErrorHandlingTest {
             // Check the stdout for the lines generated by byteman
             assertEquals(3, output.filter { it.contains("Byteman test - discharging") }.size)
             assertEquals(0, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = charlieClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
+            val (discharge, observation) = charlieClient.startFlow(StatemachineErrorHandlingTest::GetHospitalCountersFlow).returnValue.get()
             assertEquals(3, discharge)
             assertEquals(0, observation)
             assertEquals(0, aliceClient.stateMachinesSnapshot().size)
             assertEquals(0, charlieClient.stateMachinesSnapshot().size)
             // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
+            assertEquals(1, aliceClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
             // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, charlieClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
-        }
-    }
-
-    /**
-     * Throws an exception when recoding a transaction inside of [ReceiveFinalityFlow] on the responding
-     * flow's node.
-     *
-     * The flow is kept in for observation.
-     *
-     * Only the responding node keeps a checkpoint. The initiating flow has completed successfully as it has complete its
-     * send to the responding node and the responding node successfully received it.
-     */
-    @Test
-    fun `error recording a transaction inside of ReceiveFinalityFlow will keep the flow in for observation` () {
-        startDriver(notarySpec = NotarySpec(DUMMY_NOTARY_NAME, validating = false)) {
-            val charlie = createBytemanNode(CHARLIE_NAME, FINANCE_CORDAPPS)
-            val alice = createNode(ALICE_NAME, FINANCE_CORDAPPS)
-
-            // could not get rule for FinalityDoctor + observation counter to work
-            val rules = """
-                RULE Set flag when entering receive finality flow
-                CLASS ${ReceiveFinalityFlow::class.java.name}
-                METHOD call
-                AT ENTRY
-                IF !flagged("finality_flag")
-                DO flag("finality_flag"); traceln("Setting finality flag")
-                ENDRULE
-                
-                RULE Set flag when leaving resolve transactions flow
-                CLASS ${ResolveTransactionsFlow::class.java.name}
-                METHOD call
-                AT EXIT
-                IF !flagged("resolve_tx_flag")
-                DO flag("resolve_tx_flag"); traceln("Setting resolve tx flag")
-                ENDRULE
-
-                RULE Throw exception when recording transaction
-                INTERFACE ${ServiceHubInternal::class.java.name}
-                METHOD recordTransactions
-                AT ENTRY
-                IF flagged("finality_flag") && flagged("resolve_tx_flag")
-                DO traceln("Throwing exception"); 
-                    throw new java.lang.RuntimeException("die dammit die")
-                ENDRULE
-            """.trimIndent()
-
-            submitBytemanRules(rules)
-
-            val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
-            val charlieClient =
-                    CordaRPCClient(charlie.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
-
-            aliceClient.startFlow(
-                    ::CashIssueAndPaymentFlow,
-                    500.DOLLARS,
-                    OpaqueBytes.of(0x01),
-                    charlie.nodeInfo.singleIdentity(),
-                    false,
-                    defaultNotaryIdentity
-            ).returnValue.getOrThrow(30.seconds)
-
-            val (discharge, observation) = charlieClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
-            assertEquals(0, discharge)
-            assertEquals(1, observation)
-            assertEquals(0, aliceClient.stateMachinesSnapshot().size)
-            assertEquals(1, charlieClient.stateMachinesSnapshot().size)
-            // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
-            // 1 ReceiveFinalityFlow and 1 for GetNumberOfCheckpointsFlow
-            assertEquals(2, charlieClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
-        }
-    }
-
-    /**
-     * Throws an exception when resolving a transaction's dependencies inside of [ReceiveFinalityFlow] on the responding
-     * flow's node.
-     *
-     * The flow is kept in for observation.
-     *
-     * Only the responding node keeps a checkpoint. The initiating flow has completed successfully as it has complete its
-     * send to the responding node and the responding node successfully received it.
-     */
-    @Test
-    fun `error resolving a transaction's dependencies inside of ReceiveFinalityFlow will keep the flow in for observation`() {
-        startDriver(notarySpec = NotarySpec(DUMMY_NOTARY_NAME, validating = false)) {
-            val charlie = createBytemanNode(CHARLIE_NAME, FINANCE_CORDAPPS)
-            val alice = createNode(ALICE_NAME, FINANCE_CORDAPPS)
-
-            // could not get rule for FinalityDoctor + observation counter to work
-            val rules = """
-                RULE Set flag when entering receive finality flow
-                CLASS ${ReceiveFinalityFlow::class.java.name}
-                METHOD call
-                AT ENTRY
-                IF !flagged("finality_flag")
-                DO flag("finality_flag"); traceln("Setting finality flag")
-                ENDRULE
-                
-                RULE Set flag when entering resolve transactions flow
-                CLASS ${ResolveTransactionsFlow::class.java.name}
-                METHOD call
-                AT ENTRY
-                IF !flagged("resolve_tx_flag")
-                DO flag("resolve_tx_flag"); traceln("Setting resolve tx flag")
-                ENDRULE
-
-                RULE Throw exception when recording transaction
-                INTERFACE ${ServiceHubInternal::class.java.name}
-                METHOD recordTransactions
-                AT ENTRY
-                IF flagged("finality_flag") && flagged("resolve_tx_flag")
-                DO traceln("Throwing exception"); 
-                    throw new java.lang.RuntimeException("die dammit die")
-                ENDRULE
-            """.trimIndent()
-
-            submitBytemanRules(rules)
-
-            val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
-            val charlieClient =
-                    CordaRPCClient(charlie.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
-
-            aliceClient.startFlow(
-                    ::CashIssueAndPaymentFlow,
-                    500.DOLLARS,
-                    OpaqueBytes.of(0x01),
-                    charlie.nodeInfo.singleIdentity(),
-                    false,
-                    defaultNotaryIdentity
-            ).returnValue.getOrThrow(30.seconds)
-
-            val (discharge, observation) = charlieClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
-            assertEquals(0, discharge)
-            assertEquals(1, observation)
-            assertEquals(0, aliceClient.stateMachinesSnapshot().size)
-            assertEquals(1, charlieClient.stateMachinesSnapshot().size)
-            // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
-            // 1 for ReceiveFinalityFlow and 1 for GetNumberOfCheckpointsFlow
-            assertEquals(2, charlieClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
-        }
-    }
-
-    /**
-     * Throws an exception when executing [Action.CommitTransaction] as part of receiving a transaction to record inside of [ReceiveFinalityFlow] on the responding
-     * flow's node.
-     *
-     * The exception is thrown 5 times.
-     *
-     * The responding flow is retried 3 times and then completes successfully.
-     *
-     * The [StaffedFlowHospital.TransitionErrorGeneralPractitioner] catches these errors instead of the [StaffedFlowHospital.FinalityDoctor]. Due to this, the
-     * flow is retried instead of moving straight to observation.
-     */
-    @Test
-    fun `error during transition with CommitTransaction action while receiving a transaction inside of ReceiveFinalityFlow will be retried and complete successfully`() {
-        startDriver(notarySpec = NotarySpec(DUMMY_NOTARY_NAME, validating = false)) {
-            val charlie = createBytemanNode(CHARLIE_NAME, FINANCE_CORDAPPS)
-            val alice = createNode(ALICE_NAME, FINANCE_CORDAPPS)
-
-            val rules = """
-                RULE Create Counter
-                CLASS ${ActionExecutorImpl::class.java.name}
-                METHOD executeCommitTransaction
-                AT ENTRY
-                IF createCounter("counter", $counter)
-                DO traceln("Counter created")
-                ENDRULE
-                
-                RULE Set flag when entering receive finality flow
-                CLASS ${ReceiveFinalityFlow::class.java.name}
-                METHOD call
-                AT ENTRY
-                IF !flagged("finality_flag")
-                DO flag("finality_flag"); traceln("Setting finality flag")
-                ENDRULE
-                
-                RULE Throw exception on executeCommitTransaction action
-                CLASS ${ActionExecutorImpl::class.java.name}
-                METHOD executeCommitTransaction
-                AT ENTRY
-                IF flagged("finality_flag") && readCounter("counter") < 5
-                DO incrementCounter("counter"); traceln("Throwing exception"); throw new java.lang.RuntimeException("die dammit die")
-                ENDRULE
-                
-                RULE Increment discharge counter
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT READ DISCHARGE
-                IF true
-                DO traceln("Byteman test - discharging")
-                ENDRULE
-                
-                RULE Increment observation counter
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT READ OVERNIGHT_OBSERVATION
-                IF true
-                DO traceln("Byteman test - overnight observation")
-                ENDRULE
-            """.trimIndent()
-
-            submitBytemanRules(rules)
-
-            val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
-            val charlieClient =
-                    CordaRPCClient(charlie.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
-
-            aliceClient.startFlow(
-                    ::CashIssueAndPaymentFlow,
-                    500.DOLLARS,
-                    OpaqueBytes.of(0x01),
-                    charlie.nodeInfo.singleIdentity(),
-                    false,
-                    defaultNotaryIdentity
-            ).returnValue.getOrThrow(30.seconds)
-
-            val output = getBytemanOutput(charlie)
-
-            // Check the stdout for the lines generated by byteman
-            assertEquals(3, output.filter { it.contains("Byteman test - discharging") }.size)
-            assertEquals(0, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = charlieClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
-            assertEquals(3, discharge)
-            assertEquals(0, observation)
-            assertEquals(0, aliceClient.stateMachinesSnapshot().size)
-            assertEquals(0, charlieClient.stateMachinesSnapshot().size)
-            // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
-            // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, charlieClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
-        }
-    }
-
-    /**
-     * Throws an exception when executing [Action.CommitTransaction] as part of receiving a transaction to record inside of [ReceiveFinalityFlow] on the responding
-     * flow's node.
-     *
-     * The exception is thrown 7 times.
-     *
-     * The responding flow is retried 3 times and is then kept in for observation.
-     *
-     * Both the initiating node and the responding node keep checkpoints for their flows. The initiating node keeps a checkpoint for the original flow that is
-     * waiting for the responding flow's receive to complete. The responding flow's checkpoint is kept due to it failing the commit as part of receive.
-     *
-     * The [StaffedFlowHospital.TransitionErrorGeneralPractitioner] catches these errors instead of the [StaffedFlowHospital.FinalityDoctor]. Due to this, the
-     * flow is retried instead of moving straight to observation.
-     */
-    @Test
-    fun `error during transition with CommitTransaction action while receiving a transaction inside of ReceiveFinalityFlow will be retried and be kept for observation is error persists`() {
-        startDriver(notarySpec = NotarySpec(DUMMY_NOTARY_NAME, validating = false)) {
-            val charlie = createBytemanNode(CHARLIE_NAME, FINANCE_CORDAPPS)
-            val alice = createNode(ALICE_NAME, FINANCE_CORDAPPS)
-
-            val rules = """
-                RULE Create Counter
-                CLASS ${ActionExecutorImpl::class.java.name}
-                METHOD executeCommitTransaction
-                AT ENTRY
-                IF createCounter("counter", $counter)
-                DO traceln("Counter created")
-                ENDRULE
-                
-                RULE Set flag when entering receive finality flow
-                CLASS ${ReceiveFinalityFlow::class.java.name}
-                METHOD call
-                AT ENTRY
-                IF !flagged("finality_flag")
-                DO flag("finality_flag"); traceln("Setting finality flag")
-                ENDRULE
-                
-                RULE Throw exception on executeCommitTransaction action
-                CLASS ${ActionExecutorImpl::class.java.name}
-                METHOD executeCommitTransaction
-                AT ENTRY
-                IF flagged("finality_flag") && readCounter("counter") < 7
-                DO incrementCounter("counter"); traceln("Throwing exception"); throw new java.lang.RuntimeException("die dammit die")
-                ENDRULE
-                
-                RULE Increment discharge counter
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT READ DISCHARGE
-                IF true
-                DO traceln("Byteman test - discharging")
-                ENDRULE
-                
-                RULE Increment observation counter
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT READ OVERNIGHT_OBSERVATION
-                IF true
-                DO traceln("Byteman test - overnight observation")
-                ENDRULE
-            """.trimIndent()
-
-            submitBytemanRules(rules)
-
-            val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
-            val charlieClient =
-                    CordaRPCClient(charlie.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
-
-            assertFailsWith<TimeoutException> {
-                aliceClient.startFlow(
-                        ::CashIssueAndPaymentFlow,
-                        500.DOLLARS,
-                        OpaqueBytes.of(0x01),
-                        charlie.nodeInfo.singleIdentity(),
-                        false,
-                        defaultNotaryIdentity
-                ).returnValue.getOrThrow(30.seconds)
-            }
-
-            val output = getBytemanOutput(charlie)
-
-            // Check the stdout for the lines generated by byteman
-            assertEquals(3, output.filter { it.contains("Byteman test - discharging") }.size)
-            assertEquals(1, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val (discharge, observation) = charlieClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
-            assertEquals(3, discharge)
-            assertEquals(1, observation)
-            assertEquals(1, aliceClient.stateMachinesSnapshot().size)
-            assertEquals(1, charlieClient.stateMachinesSnapshot().size)
-            // 1 for CashIssueAndPaymentFlow and 1 for GetNumberOfCheckpointsFlow
-            assertEquals(2, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
-            // 1 for ReceiveFinalityFlow and 1 for GetNumberOfCheckpointsFlow
-            assertEquals(2, charlieClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
-        }
-    }
-
-    /**
-     * Triggers `killFlow` while the flow is suspended causing a [InterruptedException] to be thrown and passed through the hospital.
-     *
-     * The flow terminates and is not retried.
-     *
-     * No pass through the hospital is recorded. As the flow is marked as `isRemoved`.
-     */
-    @Test
-    fun `error during transition due to an InterruptedException (killFlow) will terminate the flow`() {
-        startDriver {
-            val alice = createBytemanNode(ALICE_NAME)
-
-            val rules = """
-                RULE Entering internal error staff member
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT ENTRY
-                IF true
-                DO traceln("Reached internal transition error staff member")
-                ENDRULE
-
-                RULE Increment discharge counter
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT READ DISCHARGE
-                IF true
-                DO traceln("Byteman test - discharging")
-                ENDRULE
-                
-                RULE Increment observation counter
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT READ OVERNIGHT_OBSERVATION
-                IF true
-                DO traceln("Byteman test - overnight observation")
-                ENDRULE
-                
-                RULE Increment terminal counter
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT READ TERMINAL
-                IF true
-                DO traceln("Byteman test - terminal")
-                ENDRULE
-            """.trimIndent()
-
-            submitBytemanRules(rules)
-
-            val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
-
-            val flow = aliceClient.startTrackedFlow(::SleepFlow)
-
-            var flowKilled = false
-            flow.progress.subscribe {
-                if (it == SleepFlow.STARTED.label) {
-                    Thread.sleep(5000)
-                    flowKilled = aliceClient.killFlow(flow.id)
-                }
-            }
-
-            assertFailsWith<TimeoutException> { flow.returnValue.getOrThrow(20.seconds) }
-
-            val output = getBytemanOutput(alice)
-
-            assertTrue(flowKilled)
-            // Check the stdout for the lines generated by byteman
-            assertEquals(0, output.filter { it.contains("Byteman test - discharging") }.size)
-            assertEquals(0, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val numberOfTerminalDiagnoses = output.filter { it.contains("Byteman test - terminal") }.size
-            assertEquals(1, numberOfTerminalDiagnoses)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
-            assertEquals(0, discharge)
-            assertEquals(0, observation)
-            assertEquals(0, aliceClient.stateMachinesSnapshot().size)
-            // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
-        }
-    }
-
-    /**
-     * Triggers `killFlow` during user application code.
-     *
-     * The user application code is mimicked by a [Thread.sleep] which is importantly not placed inside the [Suspendable]
-     * call function. Placing it inside a [Suspendable] function causes quasar to behave unexpectedly.
-     *
-     * Although the call to kill the flow is made during user application code. It will not be removed / stop processing
-     * until the next suspension point is reached within the flow.
-     *
-     * The flow terminates and is not retried.
-     *
-     * No pass through the hospital is recorded. As the flow is marked as `isRemoved`.
-     */
-    @Test
-    fun `flow killed during user code execution stops and removes the flow correctly`() {
-        startDriver {
-            val alice = createBytemanNode(ALICE_NAME)
-
-            val rules = """
-                RULE Entering internal error staff member
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT ENTRY
-                IF true
-                DO traceln("Reached internal transition error staff member")
-                ENDRULE
-
-                RULE Increment discharge counter
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT READ DISCHARGE
-                IF true
-                DO traceln("Byteman test - discharging")
-                ENDRULE
-                
-                RULE Increment observation counter
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT READ OVERNIGHT_OBSERVATION
-                IF true
-                DO traceln("Byteman test - overnight observation")
-                ENDRULE
-                
-                RULE Increment terminal counter
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT READ TERMINAL
-                IF true
-                DO traceln("Byteman test - terminal")
-                ENDRULE
-            """.trimIndent()
-
-            submitBytemanRules(rules)
-
-            val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
-
-            val flow = aliceClient.startTrackedFlow(::ThreadSleepFlow)
-
-            var flowKilled = false
-            flow.progress.subscribe {
-                if (it == ThreadSleepFlow.STARTED.label) {
-                    Thread.sleep(5000)
-                    flowKilled = aliceClient.killFlow(flow.id)
-                }
-            }
-
-            assertFailsWith<TimeoutException> { flow.returnValue.getOrThrow(30.seconds) }
-
-            val output = getBytemanOutput(alice)
-
-            assertTrue(flowKilled)
-            // Check the stdout for the lines generated by byteman
-            assertEquals(0, output.filter { it.contains("Byteman test - discharging") }.size)
-            assertEquals(0, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val numberOfTerminalDiagnoses = output.filter { it.contains("Byteman test - terminal") }.size
-            println(numberOfTerminalDiagnoses)
-            assertEquals(0, numberOfTerminalDiagnoses)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
-            assertEquals(0, discharge)
-            assertEquals(0, observation)
-            assertEquals(0, aliceClient.stateMachinesSnapshot().size)
-            // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
-        }
-    }
-
-    /**
-     * Triggers `killFlow` after the flow has already been sent to observation. The flow is not running at this point and
-     * all that remains is its checkpoint in the database.
-     *
-     * The flow terminates and is not retried.
-     *
-     * Killing the flow does not lead to any passes through the hospital. All the recorded passes through the hospital are
-     * from the original flow that was put in for observation.
-     */
-    @Test
-    fun `flow killed when it is in the flow hospital for observation is removed correctly`() {
-        startDriver {
-            val alice = createBytemanNode(ALICE_NAME)
-            val charlie = createNode(CHARLIE_NAME)
-
-            val rules = """
-                RULE Create Counter
-                CLASS ${ActionExecutorImpl::class.java.name}
-                METHOD executeSendInitial
-                AT ENTRY
-                IF createCounter("counter", $counter)
-                DO traceln("Counter created")
-                ENDRULE
-
-                RULE Throw exception on executeSendInitial action
-                CLASS ${ActionExecutorImpl::class.java.name}
-                METHOD executeSendInitial
-                AT ENTRY
-                IF readCounter("counter") < 4
-                DO incrementCounter("counter"); traceln("Throwing exception"); throw new java.lang.RuntimeException("die dammit die")
-                ENDRULE
-                
-                RULE Entering internal error staff member
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT ENTRY
-                IF true
-                DO traceln("Reached internal transition error staff member")
-                ENDRULE
-
-                RULE Increment discharge counter
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT READ DISCHARGE
-                IF true
-                DO traceln("Byteman test - discharging")
-                ENDRULE
-                
-                RULE Increment observation counter
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT READ OVERNIGHT_OBSERVATION
-                IF true
-                DO traceln("Byteman test - overnight observation")
-                ENDRULE
-                
-                RULE Increment terminal counter
-                CLASS ${StaffedFlowHospital.TransitionErrorGeneralPractitioner::class.java.name}
-                METHOD consult
-                AT READ TERMINAL
-                IF true
-                DO traceln("Byteman test - terminal")
-                ENDRULE
-            """.trimIndent()
-
-            submitBytemanRules(rules)
-
-            val aliceClient =
-                    CordaRPCClient(alice.rpcAddress).start(rpcUser.username, rpcUser.password).proxy
-
-            val flow = aliceClient.startFlow(::SendAMessageFlow, charlie.nodeInfo.singleIdentity())
-
-            assertFailsWith<TimeoutException> { flow.returnValue.getOrThrow(20.seconds) }
-
-            aliceClient.killFlow(flow.id)
-
-            val output = getBytemanOutput(alice)
-
-            // Check the stdout for the lines generated by byteman
-            assertEquals(3, output.filter { it.contains("Byteman test - discharging") }.size)
-            assertEquals(1, output.filter { it.contains("Byteman test - overnight observation") }.size)
-            val numberOfTerminalDiagnoses = output.filter { it.contains("Byteman test - terminal") }.size
-            assertEquals(0, numberOfTerminalDiagnoses)
-            val (discharge, observation) = aliceClient.startFlow(::GetHospitalCountersFlow).returnValue.get()
-            assertEquals(3, discharge)
-            assertEquals(1, observation)
-            assertEquals(0, aliceClient.stateMachinesSnapshot().size)
-            // 1 for GetNumberOfCheckpointsFlow
-            assertEquals(1, aliceClient.startFlow(::GetNumberOfCheckpointsFlow).returnValue.get())
-        }
-    }
-
-    private fun startDriver(notarySpec: NotarySpec = NotarySpec(DUMMY_NOTARY_NAME), dsl: DriverDSL.() -> Unit) {
-        driver(
-                DriverParameters(
-                        notarySpecs = listOf(notarySpec),
-                        startNodesInProcess = false,
-                        inMemoryDB = false,
-                        systemProperties = mapOf("co.paralleluniverse.fibers.verifyInstrumentation" to "true")
-                )
-        ) {
-            dsl()
-        }
-    }
-
-    private fun DriverDSL.createBytemanNode(
-            providedName: CordaX500Name,
-            additionalCordapps: Collection<TestCordapp> = emptyList()
-    ): NodeHandle {
-        return (this as InternalDriverDSL).startNode(
-                NodeParameters(
-                        providedName = providedName,
-                        rpcUsers = listOf(rpcUser),
-                        additionalCordapps = additionalCordapps
-                ),
-                bytemanPort = 12000
-        ).getOrThrow()
-    }
-
-    private fun DriverDSL.createNode(providedName: CordaX500Name, additionalCordapps: Collection<TestCordapp> = emptyList()): NodeHandle {
-        return startNode(
-                NodeParameters(
-                        providedName = providedName,
-                        rpcUsers = listOf(rpcUser),
-                        additionalCordapps = additionalCordapps
-                )
-        ).getOrThrow()
-    }
-
-    private fun submitBytemanRules(rules: String) {
-        val submit = Submit("localhost", 12000)
-        submit.addScripts(listOf(ScriptText("Test script", rules)))
-    }
-
-    private fun getBytemanOutput(nodeHandle: NodeHandle): List<String> {
-        return nodeHandle.baseDirectory
-                .list()
-                .first { it.toString().contains("net.corda.node.Corda") && it.toString().contains("stdout.log") }
-                .readAllLines()
-    }
-}
-
-@StartableByRPC
-@InitiatingFlow
-class SendAMessageFlow(private val party: Party) : FlowLogic<String>() {
-    @Suspendable
-    override fun call(): String {
-        val session = initiateFlow(party)
-        session.send("hello there")
-        return "Finished executing test flow - ${this.runId}"
-    }
-}
-
-@InitiatedBy(SendAMessageFlow::class)
-class SendAMessageResponder(private val session: FlowSession) : FlowLogic<Unit>() {
-    @Suspendable
-    override fun call() {
-        session.receive<String>().unwrap { it }
-    }
-}
-
-@StartableByRPC
-class SleepFlow : FlowLogic<Unit>() {
-
-    object STARTED : ProgressTracker.Step("I am ready to die")
-
-    override val progressTracker = ProgressTracker(STARTED)
-
-    @Suspendable
-    override fun call() {
-        sleep(Duration.of(1, ChronoUnit.SECONDS))
-        progressTracker.currentStep = STARTED
-        sleep(Duration.of(2, ChronoUnit.MINUTES))
-    }
-}
-
-@StartableByRPC
-class ThreadSleepFlow : FlowLogic<Unit>() {
-
-    object STARTED : ProgressTracker.Step("I am ready to die")
-
-    override val progressTracker = ProgressTracker(STARTED)
-
-    @Suspendable
-    override fun call() {
-        sleep(Duration.of(1, ChronoUnit.SECONDS))
-        progressTracker.currentStep = STARTED
-        logger.info("Starting ${ThreadSleepFlow::class.qualifiedName} application sleep")
-        sleep()
-        logger.info("Finished ${ThreadSleepFlow::class.qualifiedName} application sleep")
-        sleep(Duration.of(2, ChronoUnit.MINUTES))
-    }
-
-    // Sleep is moved outside of `@Suspendable` function to prevent issues with Quasar
-    private fun sleep() {
-        Thread.sleep(20000)
-    }
-}
-
-@StartableByRPC
-class GetNumberOfCheckpointsFlow : FlowLogic<Long>() {
-    override fun call(): Long {
-        return serviceHub.jdbcSession().prepareStatement("select count(*) from node_checkpoints").use { ps ->
-            ps.executeQuery().use { rs ->
-                rs.next()
-                rs.getLong(1)
-            }
-        }
-    }
-}
-
-@StartableByRPC
-class GetHospitalCountersFlow : FlowLogic<HospitalCounts>() {
-    override fun call(): HospitalCounts = HospitalCounts(
-        serviceHub.cordaService(HospitalCounter::class.java).dischargeCounter,
-        serviceHub.cordaService(HospitalCounter::class.java).observationCounter
-    )
-}
-
-@CordaSerializable
-data class HospitalCounts(val discharge: Int, val observation: Int)
-
-@Suppress("UNUSED_PARAMETER")
-@CordaService
-class HospitalCounter(services: AppServiceHub) : SingletonSerializeAsToken() {
-    var observationCounter: Int = 0
-    var dischargeCounter: Int = 0
-
-    init {
-        StaffedFlowHospital.onFlowDischarged.add { _, _ ->
-            ++dischargeCounter
-        }
-        StaffedFlowHospital.onFlowKeptForOvernightObservation.add { _, _ ->
-            ++observationCounter
+            assertEquals(1, charlieClient.startFlow(StatemachineErrorHandlingTest::GetNumberOfCheckpointsFlow).returnValue.get())
         }
     }
 }
